pyproject.toml
[project]
name = "chess_in_discord"
version = "0.1.0"
description = "Chess in discord"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "python-chess>=1.999",
    "uvicorn>=0.41.0",
    "cairosvg>=2.7.0",
]
